requirements.txt
google-genai>=1.55.0
python-dotenv
Pillow
pdf2image
python-pptx
python-docx
markdown
streamlit
